requirements.txt
opencv-python>=4.5.5
requests>=2.28.1
Pillow>=9.2.0
PyQt6>=6.3.1
numpy>=1.23.2
Shapely>=1.8.4